implemented a MovableCharacter component to make the functionality replicable across characters
